+++ custom_components/carlinko/common/consts.py
@@ -61,6 +61,7 @@ PLATFORMS = (
     "switch",
     "select",
     "device_tracker",
+    "image",
 )
 
 STORAGE_VERSION = 1